requirements.txt
pytest==5.3.2
selenium==3.14.0
allure-pytest==2.8.6
allure-python-commons==2.8.6
